backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
torch==2.7.1
torchvision==0.22.1
pillow==11.2.1
pandas==2.3.0
# scikit-learn==1.7.0
scikit-learn==1.3.2
numpy==2.2.5
openpyxl  
email-to  
